Compute marginal likelihood of a 'JAGS' model

A wrapper around bridge_sampler that automatically computes likelihood part dependent on the prior distribution and prepares parameter samples. log_posterior must specify a function that takes two arguments - a named list of samples from the prior distributions and the data, and returns log likelihood of the model part.

JAGS_bridgesampling(
  fit,
  data,
  prior_list,
  log_posterior,
  add_parameters = NULL,
  add_bounds = NULL,
  maxiter = 10000,
  silent = TRUE,
  ...
)

Arguments

fit
model fitted with either runjags posterior samples obtained with rjags-package

data
data that were used to fit the model

prior_list
named list of prior distribution (names correspond to the parameter names)

log_posterior
function that takes a named list of samples, the data, and additional list of parameters passed as ... as input and returns the log of the unnormalized posterior density of the model part

add_parameters
vector of additional parameter names that should be used in bridgesampling but were not specified in the prior_list

add_bounds
list with two name vectors ("lb" and "up") containing lower and upper bounds of the additional parameters that were not specified in the prior_list

maxiter
maximum number of iterations for the bridge_sampler

silent
whether the progress should be printed, defaults to TRUE

...
additional argument to the bridge_sampler and log_posterior function

Value

JAGS_bridgesampling returns an object of class 'bridge'.

Examples

# simulate data
set.seed(1)
data <- list(
  x = rnorm(10),
  N = 10
)
data$x

#>  [1] -0.6264538  0.1836433 -0.8356286  1.5952808  0.3295078 -0.8204684
#>  [7]  0.4874291  0.7383247  0.5757814 -0.3053884

# define priors
priors_list <- list(mu = prior("normal", list(0, 1)))

# define likelihood for the data
model_syntax <-
  "model{
    for(i in 1:N){
      x[i] ~ dnorm(mu, 1)
    }
  }"

# fit the models
fit <- JAGS_fit(model_syntax, data, priors_list, chains = 1, sample = 1000, seed = 0)

#> Loading required namespace: rjags

# define log posterior for bridge sampling
log_posterior <- function(parameters, data){
  sum(dnorm(data$x, parameters$mu, 1, log = TRUE))
}

# get marginal likelihoods
marglik <- JAGS_bridgesampling(fit, data, priors_list, log_posterior)