Add prior object to a plot

Add prior object to a plot

# S3 method for prior
lines(
  x,
  xlim = NULL,
  x_seq = NULL,
  x_range_quant = NULL,
  n_points = 1000,
  n_samples = 10000,
  force_samples = FALSE,
  transformation = NULL,
  transformation_arguments = NULL,
  transformation_settings = FALSE,
  show_parameter = if (individual) 1 else NULL,
  individual = FALSE,
  rescale_x = FALSE,
  scale_y2 = 1,
  ...
)

Arguments

x
a prior

xlim
plotting range of the prior

x_seq
sequence of x coordinates

x_range_quant
quantile used for automatically obtaining x_range if both x_range and x_seq are unspecified. Defaults to 0.005 for all but Cauchy, Student-t, Gamma, and Inverse-gamme distributions that use 0.010.

n_points
number of equally spaced points in the x_range if x_seq is unspecified

n_samples
number of samples from the prior distribution if the density cannot be obtained analytically (or if samples are forced with force_samples = TRUE)

force_samples
should prior be sampled instead of obtaining analytic solution whenever possible

transformation
transformation to be applied to the prior distribution. Either a character specifying one of the prepared transformations:

lin

linear transformation in form of a + b*x

tanh

also known as Fisher's z transformation

exp

exponential transformation

, or a list containing the transformation function fun, inverse transformation function inv, and the Jacobian of the transformation jac. See examples for details.

transformation_arguments
a list with named arguments for the transformation

transformation_settings
boolean indicating whether the settings the x_seq or x_range was specified on the transformed support

show_parameter
which parameter should be returned in case of multiple parameters per prior. Useful when priors for the omega parameter are plotted and individual = TRUE.

individual
should individual densities be returned (e.g., in case of weightfunction)

rescale_x
allows to rescale x-axis in case a weightfunction is plotted.

scale_y2
scaling factor for a secondary axis

...
additional arguments

Value

lines.prior returns NULL.

See also