Create BayesTools model tables

Creates model summary based on a model objects or provides estimates table for a runjags fit.

model_summary_table(
  model,
  model_description = NULL,
  title = NULL,
  footnotes = NULL,
  warnings = NULL,
  remove_spike_0 = TRUE,
  short_name = FALSE
)

runjags_estimates_table(
  fit,
  prior_list,
  transformations = NULL,
  title = NULL,
  footnotes = NULL,
  warnings = NULL,
  remove_spike_0 = TRUE
)

runjags_estimates_empty_table(title = NULL, footnotes = NULL, warnings = NULL)

Arguments

model
model object containing a list of priors and inference object, The inference must be a named list with information about the model: model number m_number, marginal likelihood marglik, prior and posterior probability prior_prob and post_prob, and model inclusion Bayes factor inclusion_BF

model_description
named list with additional description to be added to the table

title
title to be added to the table

footnotes
footnotes to be added to the table

warnings
warnings to be added to the table

remove_spike_0
whether prior distributions equal to spike at 0 should be removed from the prior_list

short_name
whether the prior distribution names should be shortened. Defaults to FALSE.

fit
runjags model fit

prior_list
list of prior distributions

transformations
named list of transformations to be applied to specific parameters

Value

model_summary_table returns a table with overview of the fitted model, runjags_estimates_table returns a table with MCMC estimates, and runjags_estimates_empty_table returns an empty estimates table. All of the tables are objects of class 'BayesTools_table'.

See also