Create BayesTools model tables

Creates model summary based on a model objects or provides estimates table for a runjags fit.

model_summary_table(
  model,
  model_description = NULL,
  title = NULL,
  footnotes = NULL,
  warnings = NULL,
  remove_spike_0 = TRUE,
  short_name = FALSE,
  formula_prefix = TRUE
)

runjags_estimates_table(
  fit,
  transformations = NULL,
  title = NULL,
  footnotes = NULL,
  warnings = NULL,
  remove_spike_0 = TRUE,
  transform_orthonormal = FALSE,
  formula_prefix = TRUE
)

runjags_estimates_empty_table(title = NULL, footnotes = NULL, warnings = NULL)

Arguments

model

model object containing a list of priors and inference object, The inference must be a named list with information about the model: model number m_number, marginal likelihood marglik, prior and posterior probability prior_prob and post_prob, and model inclusion Bayes factor inclusion_BF

model_description

named list with additional description to be added to the table

title

title to be added to the table

footnotes

footnotes to be added to the table

warnings

warnings to be added to the table

remove_spike_0

whether prior distributions equal to spike at 0 should be removed from the prior_list

short_name

whether the prior distribution names should be shortened. Defaults to FALSE.

formula_prefix

whether the parameter prefix from formula should be printed. Defaults to TRUE.

fit

runjags model fit

transformations

named list of transformations to be applied to specific parameters

transform_orthonormal

whether factors with orthonormal prior distributions should be transformed to differences from the grand mean

Value

model_summary_table returns a table with overview of the fitted model, runjags_estimates_table returns a table with MCMC estimates, and runjags_estimates_empty_tablereturns an empty estimates table. All of the tables are objects of class 'BayesTools_table'.

See also