Check and list 'JAGS' fitting settings

Checks and lists settings for the JAGS_fit function.

JAGS_check_and_list_fit_settings(
  chains,
  adapt,
  burnin,
  sample,
  thin,
  autofit,
  parallel,
  cores,
  silent,
  seed,
  check_mins = list(chains = 1, adapt = 50, burnin = 50, sample = 100, thin = 1),
  call = ""
)

JAGS_check_and_list_autofit_settings(
  autofit_control,
  skip_sample_extend = FALSE,
  call = ""
)

Arguments

chains
number of chains to be run, defaults to 4

adapt
number of samples used for adapting the MCMC chains, defaults to 500

burnin
number of burnin iterations of the MCMC chains, defaults to 1000

sample
number of sampling iterations of the MCMC chains, defaults to 4000

thin
thinning interval for the MCMC samples, defaults to 1

autofit
whether the models should be refitted until convergence criteria specified in autofit_control. Defaults to FALSE.

parallel
whether the chains should be run in parallel FALSE

cores
number of cores used for multithreading if parallel = TRUE, defaults to chains

silent
whether the function should proceed silently, defaults to TRUE

seed
seed for random number generation

check_mins
named list of minimal values for which should some input be checked. Defaults to:

chains

1

adapt

50

burnin

50

sample

100

thin

1

call
string to be placed as a prefix to the error call.

autofit_control
a list of arguments controlling the autofit function. Possible options are:

max_Rhat

maximum R-hat error for the autofit function. Defaults to 1.05.

min_ESS

minimum effective sample size. Defaults to 500.

max_error

maximum MCMC error. Defaults to 1.01.

max_SD_error

maximum MCMC error as the proportion of standard deviation of the parameters. Defaults to 0.05.

max_time

list specifying the time time and units after which the automatic fitting function is stopped. The units arguments need to correspond to units passed to difftime function.

sample_extend

number of samples between each convergence check. Defaults to 1000.

skip_sample_extend
whether sample_extend is allowed to be NULL and skipped in the check

Value

JAGS_check_and_list_fit_settings invisibly returns a list of checked fit settings. JAGS_check_and_list_autofit_settings invisibly returns a list of checked autofit settings. parameter names.