Names of medical subfields from the Cochrane database of systematic reviews

Contain names identifying the individual subfields from the Cochrane database of systematic reviews. The individual elements correspond to valid name arguments for the prior_informed() function.

prior_informed_medicine_names

Format

An object of class character of length 47.

Value

returns a character vector with names of medical subfields from Cochrane database of systematic reviews.

See also

Examples

print(prior_informed_medicine_names)
#>  [1] "Acute Respiratory Infections"                     
#>  [2] "Airways"                                          
#>  [3] "Anaesthesia"                                      
#>  [4] "Back and Neck"                                    
#>  [5] "Bone, Joint and Muscle Trauma"                    
#>  [6] "Colorectal"                                       
#>  [7] "Common Mental Disorders"                          
#>  [8] "Consumers and Communication"                      
#>  [9] "Cystic Fibrosis and Genetic Disorders"            
#> [10] "Dementia and Cognitive Improvement"               
#> [11] "Developmental, Psychosocial and Learning Problems"
#> [12] "Drugs and Alcohol"                                
#> [13] "Effective Practice and Organisation of Care"      
#> [14] "Emergency and Critical Care"                      
#> [15] "ENT"                                              
#> [16] "Eyes and Vision"                                  
#> [17] "Gynaecological, Neuro-oncology and Orphan Cancer" 
#> [18] "Gynaecology and Fertility"                        
#> [19] "Heart"                                            
#> [20] "Hepato-Biliary"                                   
#> [21] "HIV/AIDS"                                         
#> [22] "Hypertension"                                     
#> [23] "Incontinence"                                     
#> [24] "Infectious Diseases"                              
#> [25] "Inflammatory Bowel Disease"                       
#> [26] "Injuries"                                         
#> [27] "Kidney and Transplant"                            
#> [28] "Metabolic and Endocrine Disorders"                
#> [29] "Methodology"                                      
#> [30] "Movement Disorders"                               
#> [31] "Musculoskeletal"                                  
#> [32] "Neonatal"                                         
#> [33] "Oral Health"                                      
#> [34] "Pain, Palliative and Supportive Care"             
#> [35] "Pregnancy and Childbirth"                         
#> [36] "Public Health"                                    
#> [37] "Schizophrenia"                                    
#> [38] "Sexually Transmitted Infections"                  
#> [39] "Skin"                                             
#> [40] "Stroke"                                           
#> [41] "Tobacco Addiction"                                
#> [42] "Upper GI and Pancreatic Diseases"                 
#> [43] "Urology"                                          
#> [44] "Vascular"                                         
#> [45] "Work"                                             
#> [46] "Wounds"                                           
#> [47] "Cochrane"