Weight functions

Marginal density, marginal distribution function, marginal quantile function and random generation for weight functions.

mdone.sided(x, alpha = NULL, alpha1 = NULL, alpha2 = NULL, log = FALSE)

mdtwo.sided(x, alpha, log = FALSE)

mdone.sided_fixed(x, omega, log = FALSE)

mdtwo.sided_fixed(x, omega, log = FALSE)

rone.sided(n, alpha = NULL, alpha1 = NULL, alpha2 = NULL)

rtwo.sided(n, alpha)

rone.sided_fixed(n, omega)

rtwo.sided_fixed(n, omega)

mpone.sided(
  q,
  alpha = NULL,
  alpha1 = NULL,
  alpha2 = NULL,
  lower.tail = TRUE,
  log.p = FALSE
)

mptwo.sided(q, alpha, lower.tail = TRUE, log.p = FALSE)

mpone.sided_fixed(q, omega, lower.tail = TRUE, log.p = FALSE)

mptwo.sided_fixed(q, omega, lower.tail = TRUE, log.p = FALSE)

mqone.sided(
  p,
  alpha = NULL,
  alpha1 = NULL,
  alpha2 = NULL,
  lower.tail = TRUE,
  log.p = FALSE
)

mqtwo.sided(p, alpha, lower.tail = TRUE, log.p = FALSE)

mqone.sided_fixed(p, omega, lower.tail = TRUE, log.p = FALSE)

mqtwo.sided_fixed(p, omega, lower.tail = TRUE, log.p = FALSE)

Arguments

x, q

vector or matrix of quantiles.

alpha

vector or matrix with concentration parameters for the Dirichlet distribution for a monotonic one.sided or a two.sided weight function.

alpha1

vector or matrix with concentration parameters for the Dirichlet distribution for the expected direction of non-monotonic one.sided of weight function.

alpha2

vector or matrix with concentration parameters for the Dirichlet distribution for the unexpected direction of non-monotonic one.sided of weight function.

log, log.p

logical; if TRUE, probabilities p are given as log(p).

omega

vector or matrix of fixed probabilities for a one.sided or a two.sided weight function.

n

number of observations.

lower.tail

logical; if TRUE (default), probabilities are \(P[X \le x]\), otherwise, \(P[X \ge x]\).

p

vector of probabilities.

Value

mdone.sided, mdtwo.sided, mdone.sided_fixed, and mdtwo.sided_fixed give the marginal density, mpone.sided, mptwo.sided, mpone.sided_fixed, and mptwo.sided_fixed give the marginal distribution function, mqone.sided, mqtwo.sided, mqone.sided_fixed, and mqtwo.sided_fixed give the marginal quantile function, and rone.sided, rtwo.sided, rone.sided_fixed, and rtwo.sided_fixed generate random deviates.

Examples

# draw samples from a two-sided weight function
rtwo.sided(10, alpha = c(1, 1))
#>             [,1] [,2]
#>  [1,] 0.46439642    1
#>  [2,] 0.04716989    1
#>  [3,] 0.71910622    1
#>  [4,] 0.95906075    1
#>  [5,] 0.94976423    1
#>  [6,] 0.25670743    1
#>  [7,] 0.88470137    1
#>  [8,] 0.88922749    1
#>  [9,] 0.66088187    1
#> [10,] 0.72228484    1

# draw samples from a monotone one-sided weight function
rone.sided(10, alpha = c(1, 1, 1))
#>             [,1]      [,2] [,3]
#>  [1,] 0.01313900 0.7528822    1
#>  [2,] 0.39695058 0.9523203    1
#>  [3,] 0.51717076 0.6036944    1
#>  [4,] 0.32504571 0.5099114    1
#>  [5,] 0.13829948 0.7936787    1
#>  [6,] 0.05130129 0.9083622    1
#>  [7,] 0.03663619 0.3303725    1
#>  [8,] 0.50885810 0.7152161    1
#>  [9,] 0.18002423 0.7771316    1
#> [10,] 0.19877398 0.6412896    1

# draw samples from a non-monotone one-sided weight function
rone.sided(10, alpha1 = c(1, 1), alpha2 = c(1, 1))
#>            [,1]       [,2] [,3]
#>  [1,] 0.7060430 0.20216529    1
#>  [2,] 0.9672913 0.01495326    1
#>  [3,] 0.8185752 0.77588693    1
#>  [4,] 0.5598840 0.03990301    1
#>  [5,] 0.7815403 0.58610764    1
#>  [6,] 0.9984784 0.87007164    1
#>  [7,] 0.9133474 0.34747997    1
#>  [8,] 0.6409038 0.01194426    1
#>  [9,] 0.9905972 0.04908653    1
#> [10,] 0.9509986 0.20771290    1